tests/wpt/web-platform-tests/vibration/api-is-present.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset='utf-8'/>
    <title>Vibration API: test that the vibrate() method is present (with or without vendor prefix)</title>
    <link rel='author' title='Robin Berjon' href='mailto:robin@berjon.com'/>
    <link rel='help' href='http://www.w3.org/TR/vibration/#methods'/>
    <meta name='flags' content='dom'/>
    <meta name='assert' content='Check that the vibrate() method is present.'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
      This test checks for the presence of the <code>vibrate()</code> method, taking
      vendor prefixes into account.
    </p>
    <div id='log'></div>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script src='/common/vendor-prefix.js' data-prefixed-objects='[{"ancestors":["navigator"], "name":"vibrate"}]'></script>
    <script>
      test(function () {
          assert_true(undefined !== navigator.vibrate, "navigator.vibrate exists");
      }, "vibrate() is present on navigator");
    </script>
  </body>
</html>
